+++ src/app/pages/programar-mantenimiento/programar-mantenimiento-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { ProgramarMantenimientoPage } from './programar-mantenimiento.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ProgramarMantenimientoPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ProgramarMantenimientoPageRoutingModule {}